requirements.txt
fastapi
uvicorn
psycopg2-binary
sentence-transformers
langchain-text-splitters
pypdf
python-dotenv
requests
python-docx
google-generativeai